Added car.php page

+++ public_html/announcements.php
@@ -12,13 +12,22 @@ if($conn->connect_error){
 
 	#$message = htmlspecialchars($_POST['text']);
 	$message = htmlspecialchars("This worked fine!!!");
+	#$eid = $_POST['eid']; ### Event ID for the event in which the announcement is made for 
+	$eid = 1;
 
 	if (strlen($message) < 2047) {
 		$query = 'INSERT INTO Announcement (content, aDate) VALUES (?, ?);';
 		$stmt = $conn->prepare($query);
 		$stmt->bind_param("ss", $message, date("Y-m-d H:i:s"));
 		if ($stmt->execute()) {
-			echo "New announcement created successfully!!!";
+			echo "$stmt->insert_id";
+			$aid = $stmt->insert_id;
+			$stmt2 = $conn->prepare("INSERT INTO event_has_announcements (aid, eid) VALUES (?,?);");
+			$stmt2->bind_param('ii', $aid, $eid);
+
+			if ($stmt2->execute()) {
+				echo "New announcement created successfully!!!";
+			}
 		}
 		else {
 			echo "Something went wrong with insert function";
@@ -32,11 +41,11 @@ if($conn->connect_error){
 #}
 
 ##### Function to edit an announcement
-#else if ($_POST['action']=='edit'){
-	#$message = htmlspecialchars($_POST['content']);  #this post and the one below would be hidden fields
-	$message = htmlspecialchars('The old the old message'); 
-	#$announcement_key = $_POST['announcement_primary_key'];  #this would also be a hidden field 
-	$announcement_key = 3;
+else if ($_POST['action']=='edit'){
+	$message = htmlspecialchars($_POST['content']);  #this post and the one below would be hidden fields
+	#$message = htmlspecialchars('The old the old message'); 
+	$announcement_key = $_POST['announcement_primary_key'];  #this would also be a hidden field 
+	#$announcement_key = 3;
 
 	$query = 'UPDATE Announcement SET content=?, aDate=? WHERE (aid = ?);';
 	$stmt = $conn->prepare($query);
@@ -48,12 +57,12 @@ if($conn->connect_error){
 		echo "Something went wrong with the edit function";
 	}
 	$stmt->close();
-#}
+}
 
 ##### Function to remove an announcement
-#else if ($_POST['action']=='remove'){
-	#$announcement_id = htmlspecialchars($_POST['announcement_id']);
-	$announcement_id = htmlspecialchars('1');
+else if ($_POST['action']=='remove'){
+	$announcement_id = htmlspecialchars($_POST['announcement_id']);
+	#$announcement_id = htmlspecialchars('1');
 
 	$query = 'DELETE FROM Announcement WHERE (aid = ?) LIMIT 1;';
 	$stmt = $conn->prepare($query);
@@ -65,7 +74,7 @@ if($conn->connect_error){
 		echo "Something went wrong with remove function";
 	}
 	$stmt->close();
-#}
+}
 
 ##### Retrieves all announcements 
 #else if ($_POST['action']=='get'){  #This all works perfectly
